refactor: standardize namespace and coding style
This commit focuses on code refactoring to improve consistency and maintainability across the codebase:

1. Added proper namespace hierarchy with 'laminpie' as the root namespace
2. Standardized method naming convention (PascalCase for public methods)
3. Replaced old string handling (const char*) with std::string
4. Fixed type casting issues with explicit static_cast
5. Updated CMake configuration to properly include system modules
6. Refactored event system with more descriptive event type names (kEventName)
7. Replaced outdated logging macros with standardized system log macros
8. Removed deprecated framework implementation (framwork.cpp/h files cleaned)
9. Enhanced type safety with proper type aliases and consistent parameter naming

---

重构：标准化命名空间和编码风格

此次提交主要对代码进行了重构，以提高代码一致性和可维护性：

1. 添加了以'laminpie'为根的规范化命名空间层次结构
2. 统一了方法命名约定（公共方法使用PascalCase）
3. 将旧的字符串处理（const char*）替换为std::string
4. 使用明确的static_cast修复类型转换问题
5. 更新了CMake配置以正确包含系统模块
6. 重构了事件系统，使用更具描述性的事件类型名称（kEventName）
7. 将过时的日志宏替换为标准化的系统日志宏
8. 移除了废弃的框架实现（清空framwork.cpp/h文件）
9. 通过合适的类型别名和一致的参数命名增强了类型安全性

+++ src/core/gui/lvgl/laminpie_lv_display.cpp
@@ -42,7 +42,7 @@ bool LvDisplay::fontCalibrateMethod(StyleFont &target, const StyleSize *parent) 
 
     // Size
     CheckValueAndReturn(
-        target.size_px, StyleFont::FONT_SIZE_MIN, StyleFont::FONT_SIZE_MAX, false, "Invalid size"
+        static_cast<int>(target.size_px), StyleFont::FONT_SIZE_MIN, StyleFont::FONT_SIZE_MAX, false, "Invalid size"
     );
     // Font description
     if (font_resource == nullptr) {
@@ -57,10 +57,10 @@ process_height:
     // Height
     if (target.flags.enable_height_percent) {
         CheckNullAndReturn(parent, false, "Invalid parent size");
-        CheckValueAndReturn(target.height_percent, 1, 100, false, "Invalid height percent");
+        CheckValueAndReturn(static_cast<int>(target.height_percent), 1, 100, false, "Invalid height percent");
         target.height = (parent->height * target.height_percent) / 100;
     } else if (parent != nullptr) {
-        CheckValueAndReturn(target.height, 1, parent->height, false, "Invalid height");
+        CheckValueAndReturn(static_cast<int>(target.height), 1, parent->height, false, "Invalid height");
     }
 
     // Font description & size
@@ -92,7 +92,7 @@ bool LvDisplay::updateByNewData(void)
 const lv_font_t *LvDisplay::getFontBySize(uint8_t size_px) const
 {
     CheckValueAndReturn(
-        static_cast<uint8_t>(size_px), StyleFont::FONT_SIZE_MIN, StyleFont::FONT_SIZE_MAX, nullptr, "Invalid size"
+        static_cast<int>(size_px), StyleFont::FONT_SIZE_MIN, StyleFont::FONT_SIZE_MAX, nullptr, "Invalid size"
     );
 
     auto it = _size_font_map.find(size_px);
